+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - (Feature) Envoy AuthV3 Integration
 - (Maintenance) Switch to ubuntu:24.04 base image
 - (Feature) Gateway Group for ArangoDeployment
+- (Feature) Gateway config loader
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries